Support transfer object transaction type in stress tests
