+++ crates/sui-benchmark/Cargo.toml
@@ -21,10 +21,12 @@ rocksdb = "0.18.0"
 serde_with = { version = "1.14.0", features = ["hex"] }
 tracing = "0.1.35"
 tracing-subscriber = { version = "0.3.11", features = ["time", "registry", "env-filter"] }
+telemetry-subscribers = { git = "https://github.com/MystenLabs/mysten-infra", rev = "bdaf6ee818497e6772f3000f938f1e5710f6890f" }
 clap = { version = "3.1.17", features = ["derive"] }
 prometheus = "0.13.1"
 multiaddr = "0.14.0"
 crossterm = "0.23.2"
+rand = "0.8.5"
 
 bcs = "0.1.3"
 sui-core = { path = "../sui-core" }
